requirements.txt
click==8.0.0
colorama==0.4.4
pex==2.1.39
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-watch==4.2.0
pytest-cov==2.12.0
gitdb==4.0.7
GitPython==3.1.17
termcolor==1.1.0
yaspin==2.0.0
